Move TreeIterator to separate source file from rest of ts module.
